Seperate the part of preprocessing from agent
